requirements.txt
### Root Packages
#Flask==1.0.2
#python-dotenv==0.8.2
#PyMySQL==0.8.1
#Flask-SQLAlchemy
#Flask-Migrate==2.1.1
#Flask-WTF==0.14
#Flask-Markdown==0.3
#python-slugify==1.2.5
#Pillow==5.1.0
alembic==0.9.9
click==6.7
Flask==1.0.2
Flask-Markdown==0.3
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
Pillow==5.1.0
PyMySQL==0.8.1
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
python-slugify==1.2.5
six==1.11.0
SQLAlchemy==1.2.8
Unidecode==1.0.22
Werkzeug==0.14.1
WTForms==2.2.1
